squashed changes from migrate-to-metadata merge
